Moved GL logging in Startup class
